Fix mypy typecheck violations repo-wide
The "quality" CI check runs ruff then mypy sequentially; once ruff
passed (previous commit), mypy's 47 pre-existing errors across 14
files surfaced for the first time. All pre-existing debt, unrelated
to any single feature:

- no-any-return fixes where urllib/pydantic/pow() calls implicitly
  returned Any (legislative_watch, europe_pmc fetcher, coverage
  scoring, recency scoring)
- apps/api/main.py's middleware had call_next typed as returning
  Response instead of Awaitable[Response], so `await`ing it didn't
  typecheck
- supabase_auth_provider's try/except import-guard variable needed
  an explicit Optional annotation
- container.py's repository/retriever builders returned bare
  `object` instead of their Protocol port types, so every service
  constructor call downstream looked type-mismatched
- several test helper factories (`_source`, `_outcome`, `_result`)
  built untyped **kwargs dicts and unpacked them into pydantic
  models; switched to `.model_validate()` and typed the kwargs
  instead of relying on `**dict[str, object]` matching each field
- a test double (FixedInterviewPlanner) needed to actually subclass
  InterviewPlanner rather than just duck-type it, since it isn't a
  Protocol
- presidio-analyzer and presidio-anonymizer each define their own
  RecognizerResult class (structurally identical, this is how the
  two packages are meant to be wired together) — annotated with
  type: ignore[arg-type] rather than worked around

`ruff check .`, `mypy apps packages tests`, and the full test suite
are all clean.

+++ packages/core/domain/knowledge/quality.py
@@ -74,7 +74,7 @@ def _recency_score(year: int | None, *, now_year: int, half_life_years: int = 12
     if year is None:
         return 0.4  # undated evidence isn't worthless, just not freshness-scored
     age = max(0, now_year - year)
-    return 0.5 ** (age / half_life_years)
+    return float(0.5 ** (age / half_life_years))
 
 
 def _evidence_depth_score(access_depth: str) -> float:
